requirements.txt
datasets==2.13.1
poe_api==0.4.8
sounddevice==0.4.6
soundfile==0.12.1
torch==2.0.1
transformers==4.31.0.dev0
python-dotenv==1.0.0
